Generalize more tests from the JVM to other platforms
